requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
idna==3.4
loguru==0.7.0
peewee==3.16.2
pyTelegramBotAPI==4.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot-calendar==1.0.5
requests==2.31.0
six==1.16.0
urllib3==1.26.16
win32-setctime==1.1.0
